Remove traces of old composite build testing
